model WIP6

+++ iTunesSearchCacher/CoreData/ArtistEntity+CoreDataProperties.swift
@@ -14,7 +14,7 @@ import CoreData
 
 extension ArtistEntity {
 
-    @NSManaged var artistId: Int32
+    @NSManaged var artistId: Int64
     @NSManaged var artistName: String
     @NSManaged var artistViewUrl: String
     @NSManaged var collections: NSSet
